requirements.txt
Flask==3.0.3
mysql_connector_repackaged==0.3.1
openai==1.51.2
Pillow==10.4.0
redis==5.0.3
redis==5.1.1
Requests==2.32.3
